+++ NEWS.md
@@ -6,6 +6,7 @@
 
 - `<expr>$item()` to strictly extract a single value from an expression (#1652).
 - `<expr>$arr$eval()` to run any Polars expression on all subarrays of an Array column (#1653).
+- `<expr>$name$replace()` to replace expression names using regular expressions (#1654).
 
 ## polars 1.6.0